added more attack features

+++ LoserGameForLosers/FadeComponent.h
@@ -2,13 +2,13 @@
 
 class FadeComponent : public Component
 {
-	int frames_per_cycle_;
+	Uint8 frames_per_cycle_;
 	int cycles_completed_;
 	TextureComponent * texture_;
 	Uint8 alpha_;
 	bool waxing_;
 public:
-	explicit FadeComponent(int frames_per_cycle);
+	explicit FadeComponent(Uint8 frames_per_cycle);
 
 	void init() override;
 	void update() override;
